небольшие косметические правки читабельности
