+++ pom.xml
@@ -4,9 +4,9 @@
 				 xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
 	<modelVersion>4.0.0</modelVersion>
 
-	<groupId>demidko.cf</groupId>
-	<artifactId>bflexion</artifactId>
-	<version>1.0-SNAPSHOT</version>
+	<groupId>cf.demidko.aot</groupId>
+	<artifactId>aot-lematizer</artifactId>
+	<version>2.0</version>
 
 	<properties>
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
