+++ src/main/java/cf/demidko/aot/LemmasStorage.java
@@ -1,6 +1,6 @@
-package cf.demidko.bflexion;
+package cf.demidko.aot;
 
-import me.bazhenov.Utils;
+import me.bazhenov.aot.Utils;
 
 import java.io.IOException;
 import java.io.InputStream;
