+++ src/main/java/me/bazhenov/aot/Utils.java
@@ -1,4 +1,4 @@
-package me.bazhenov;
+package me.bazhenov.aot;
 
 public class Utils {
