+++ src/test/java/cf/demidko/aot/bencmarks/LoadDictionaryBenchmark.java
@@ -1,6 +1,6 @@
-package cf.demidko.bflexion.bencmarks;
+package cf.demidko.aot.bencmarks;
 
-import cf.demidko.bflexion.FlexionStorage;
+import cf.demidko.aot.FlexionStorage;
 import org.openjdk.jmh.annotations.Benchmark;
 import org.openjdk.jmh.annotations.BenchmarkMode;
 import org.openjdk.jmh.annotations.Mode;
